correction

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -11,20 +11,20 @@
 
 int main(void)
 {
-	unsigned long fib1 = 1;
-	unsigned long fib2 = 2;
+	unsigned int fib1 = 1;
+	unsigned int fib2 = 2;
 	int i;
 
-	printf("%lu, %lu", fib1, fib2);
+	printf("%u, %u", fib1, fib2);
 
 	for (i = 3; i <= 98; i++)
 	{
-		unsigned long currentfib = fib1 + fib2;
+		unsigned int currentfib = fib1 + fib2;
 
 		if (i < 98)
-			printf(", %lu", currentfib);
+			printf(", %u", currentfib);
 		else
-			printf(", %lu\n", currentfib);
+			printf(", %u\n", currentfib);
 
 		fib1 = fib2;
 		fib2 = currentfib;
